Move Tailwind config out into top-level of repo
Want to share this config between the Resume stuff and the Next.js stuff; bump it out
