+++ next-js/package.json
@@ -11,6 +11,7 @@
     "generate-location-data": "tsx app/music/scripts/generate-location-data.ts"
   },
   "dependencies": {
+    "@andrewwestling/tailwind-config": "file:../tailwind",
     "@mdx-js/loader": "^3.0.1",
     "@mdx-js/react": "^3.0.1",
     "@next/mdx": "15.1.6",
