+++ tailwind/src/globals.css
@@ -1,0 +1,121 @@
+@tailwind base;
+@tailwind components;
+@tailwind utilities;
+
+@layer base {
+  a {
+    @apply underline;
+  }
+  a:hover {
+    @apply hover:text-primary;
+  }
+
+  /* To use with MDX components to revert list styles back to normal (Tailwind resets it) */
+  ul.list-revert,
+  ol.list-revert {
+    list-style: revert;
+  }
+}
+
+@layer components {
+  /* Typography Presets */
+
+  /* Really small labels labels etc */
+  .text-preset-1 {
+    @apply text-xs font-normal leading-normal;
+  }
+
+  /* Semibold text in the context of a paragraph */
+  .text-preset-1-bold {
+    @apply text-xs font-semibold leading-normal;
+  }
+
+  /* Small labels etc */
+  .text-preset-2 {
+    @apply text-sm font-normal leading-normal;
+  }
+
+  /* Semibold text in the context of a paragraph */
+  .text-preset-2-bold {
+    @apply text-sm font-semibold leading-normal;
+  }
+
+  /* Normal text  */
+  .text-preset-3 {
+    @apply text-base font-normal leading-normal;
+  }
+
+  /* Semibold text in the context of a paragraph */
+  .text-preset-3-bold {
+    @apply text-base font-semibold leading-normal;
+  }
+
+  /* MDX <h4>, <h5>, <h6> */
+  .text-preset-4 {
+    @apply text-xl sm:text-2xl font-medium leading-snug tracking-tight;
+  }
+
+  /* Semibold text in the context of a paragraph */
+  .text-preset-4-bold {
+    @apply text-xl sm:text-2xl font-semibold leading-snug tracking-tight;
+  }
+
+  /* MDX <h3> */
+  .text-preset-5 {
+    @apply text-2xl sm:text-3xl font-semibold leading-snug tracking-tight;
+  }
+
+  /* Semibold text in the context of a paragraph */
+  .text-preset-5-bold {
+    @apply text-2xl sm:text-3xl font-bold leading-snug tracking-tight;
+  }
+
+  /* MDX <h2>, Music Library SectionHeading */
+  .text-preset-6 {
+    @apply text-3xl sm:text-4xl font-semibold leading-none tracking-tight;
+  }
+
+  /* Semibold text in the context of a paragraph */
+  .text-preset-6-bold {
+    @apply text-3xl sm:text-4xl font-bold leading-none tracking-tight;
+  }
+
+  /* MDX <h1>, Music Library PageTitle */
+  .text-preset-7 {
+    @apply text-4xl sm:text-5xl font-semibold leading-none tracking-tight;
+  }
+
+  /* Semibold text in the context of a paragraph */
+  .text-preset-7-bold {
+    @apply text-4xl sm:text-5xl font-bold leading-none tracking-tight;
+  }
+
+  /* Markdown styles */
+  .markdown {
+    @apply text-preset-3;
+  }
+
+  .markdown p {
+    @apply text-preset-3;
+  }
+
+  .markdown strong {
+    @apply font-semibold;
+  }
+
+  .markdown a {
+    @apply hover:text-primary;
+  }
+
+  .markdown ul {
+    @apply list-disc list-inside space-y-1;
+  }
+
+  .markdown li {
+    @apply text-preset-3;
+  }
+
+  .markdown li p {
+    @apply inline;
+  }
+}
